Refactor into a different file
